sync X, scale and b as same type

+++ onnx/test/shape_inference_test.py
@@ -3488,8 +3488,8 @@ class TestShapeInference(unittest.TestCase):
     def test_batch_norm_train_with_diff_type(self):  # type: () -> None
         graph = self._make_graph(
             [('x', TensorProto.FLOAT16, (3, 4, 5, 6, 7)),
-             ('scale', TensorProto.FLOAT, (4,)),
-             ('b', TensorProto.FLOAT, (4,)),
+             ('scale', TensorProto.FLOAT16, (4,)),
+             ('b', TensorProto.FLOAT16, (4,)),
              ('input_mean', TensorProto.FLOAT, (4,)),
              ('input_var', TensorProto.FLOAT, (4,))],
             [make_node('BatchNormalization', ['x', 'scale', 'b', 'input_mean', 'input_var'],
